# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MetaMart API",
    "description": "MetaMart API Information",
    "servers": [
      "http://localhost:5001"
    ]
  },
  "host": "localhost:5001",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/company/": {
      "get": {
        "tags": [
          "companiesController"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/company/add": {
      "post": {
        "tags": [
          "companiesController"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "companyName": {
                  "example": "any"
                },
                "industry": {
                  "example": "any"
                },
                "addressId": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "contactInformation": {
                  "example": "any"
                },
                "website": {
                  "example": "any"
                },
                "industryType": {
                  "example": "any"
                },
                "shippingInformation": {
                  "example": "any"
                },
                "paymentMethods": {
                  "example": "any"
                },
                "returnPolicy": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/order/": {
      "get": {
        "tags": [
          "orderController"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/order/user": {
      "get": {
        "tags": [
          "orderController"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/order/role": {
      "get": {
        "tags": [
          "orderController"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/order/create": {
      "post": {
        "tags": [
          "orderController"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "paymentMethod": {
                  "example": "any"
                },
                "trackingNumber": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "contactNumber": {
                  "example": "any"
                },
                "zipCode": {
                  "example": "any"
                },
                "additionalInfo": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "country": {
                  "example": "any"
                },
                "shippingAddress": {
                  "example": "any"
                },
                "orderItems": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/order/update": {
      "put": {
        "tags": [
          "orderController"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "paymentMethod": {
                  "example": "any"
                },
                "trackingNumber": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "contactNumber": {
                  "example": "any"
                },
                "zipCode": {
                  "example": "any"
                },
                "additionalInfo": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "country": {
                  "example": "any"
                },
                "shippingAddress": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/order/delete/{orderId}": {
      "delete": {
        "tags": [
          "orderController"
        ],
        "description": "",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}